Basic game and UI done. On splitting yet and Josh to add probability bits.

+++ api/Domain/NewGame.cs
@@ -4,6 +4,6 @@ namespace common.Domain
     {
         public int UserId { get; set; }
         public int NumberOfBots { get; set; }
-        public decimal BetPlaced { get; set; }
+        public int BetPlaced { get; set; }
     }
 }
